Move AppContext into Context.ts
