+++ src/components/Context.tsx
@@ -5,9 +5,28 @@ import { Provider } from "react-redux";
 import type { Language } from "../CodeMirrorBlocks";
 import { CMBEditor } from "../editor";
 import type { AppStore } from "../store";
+import { Search } from "../ui/BlockEditor";
 
 export const EditorContext = React.createContext<CMBEditor | null>(null);
 export const LanguageContext = React.createContext<Language | null>(null);
+
+export type AppHelpers = {
+  /**
+   * @internal
+   * Dialog showing/hiding methods deal with ToggleEditor state.
+   * We pass them to mode-specific components, to allow those
+   * components to show/hide dialogs
+   *
+   * This is hooked up when ToggleEditor gets mounted
+   */
+  showDialog?: (
+    contents: null | { title: string; content: React.ReactElement }
+  ) => void;
+  focusToolbar?: () => void;
+  search?: Search;
+};
+
+export const AppContext = React.createContext<AppHelpers>({});
 
 export default function Context(props: {
   store: AppStore;
